背景がスクロールするように変更 背景画像変更

+++ RaidenProject/CharaBase.h
@@ -3,7 +3,7 @@
 class CharaBase
 {
 private:
-	int Speed;
+	float Speed;
 	int Image;
 public:
 	CharaBase();
